profile public almost done
